requirements.txt
boto3
typer[all]
python-dotenv
